verisource-ai/backend/requirements.txt
fastapi
uvicorn
python-dotenv
loguru
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
chromadb
sentence-transformers
pypdf
groq